pyproject.toml
[project]
name = "airflow-validator"
version = "0.1.0"
description = "Airflow DAG 실행 결과 검증 도구"
requires-python = ">=3.9"
dependencies = [
    "pymysql>=1.1.0",
    "psycopg2-binary>=2.9.9",
    "httpx>=0.27.0",
    "openpyxl>=3.1.2",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
airflow-validator = "airflow_validator.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
